+++ x/btcstkconsumer/genesis.go
@@ -3,21 +3,25 @@ package btcstkconsumer
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/babylonlabs-io/babylon/x/btcstkconsumer/keeper"
-	"github.com/babylonlabs-io/babylon/x/btcstkconsumer/types"
+	"github.com/babylonlabs-io/babylon/v2/x/btcstkconsumer/keeper"
+	"github.com/babylonlabs-io/babylon/v2/x/btcstkconsumer/types"
 )
 
 // InitGenesis initializes the module's state from a provided genesis state.
 func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) {
-	if err := k.SetParams(ctx, genState.Params); err != nil {
+	if err := genState.Validate(); err != nil {
+		panic(err)
+	}
+	if err := k.InitGenesis(ctx, genState); err != nil {
 		panic(err)
 	}
 }
 
 // ExportGenesis returns the module's exported genesis.
 func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
-	genesis := types.DefaultGenesis()
-	genesis.Params = k.GetParams(ctx)
-
-	return genesis
+	gs, err := k.ExportGenesis(ctx)
+	if err != nil {
+		panic(err)
+	}
+	return gs
 }