+++ x/incentive/keeper/reward_tracker_store_test.go
@@ -9,10 +9,10 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	appparams "github.com/babylonlabs-io/babylon/app/params"
-	"github.com/babylonlabs-io/babylon/testutil/datagen"
-	"github.com/babylonlabs-io/babylon/testutil/store"
-	"github.com/babylonlabs-io/babylon/x/incentive/types"
+	appparams "github.com/babylonlabs-io/babylon/v2/app/params"
+	"github.com/babylonlabs-io/babylon/v2/testutil/datagen"
+	"github.com/babylonlabs-io/babylon/v2/testutil/store"
+	"github.com/babylonlabs-io/babylon/v2/x/incentive/types"
 )
 
 func FuzzCheckBtcDelegationActivated(f *testing.F) {