+++ x/monitor/keeper/grpc_query_test.go
@@ -8,13 +8,13 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 
-	"github.com/babylonlabs-io/babylon/app"
-	"github.com/babylonlabs-io/babylon/btctxformatter"
-	"github.com/babylonlabs-io/babylon/testutil/datagen"
-	"github.com/babylonlabs-io/babylon/testutil/mocks"
-	ckpttypes "github.com/babylonlabs-io/babylon/x/checkpointing/types"
-	types2 "github.com/babylonlabs-io/babylon/x/epoching/types"
-	"github.com/babylonlabs-io/babylon/x/monitor/types"
+	"github.com/babylonlabs-io/babylon/v2/app"
+	"github.com/babylonlabs-io/babylon/v2/btctxformatter"
+	"github.com/babylonlabs-io/babylon/v2/testutil/datagen"
+	"github.com/babylonlabs-io/babylon/v2/testutil/mocks"
+	ckpttypes "github.com/babylonlabs-io/babylon/v2/x/checkpointing/types"
+	types2 "github.com/babylonlabs-io/babylon/v2/x/epoching/types"
+	"github.com/babylonlabs-io/babylon/v2/x/monitor/types"
 )
 
 func FuzzQueryEndedEpochBtcHeight(f *testing.F) {