+++ x/monitor/module.go
@@ -12,9 +12,9 @@ import (
 
 	abci "github.com/cometbft/cometbft/abci/types"
 
-	"github.com/babylonlabs-io/babylon/x/monitor/client/cli"
-	"github.com/babylonlabs-io/babylon/x/monitor/keeper"
-	"github.com/babylonlabs-io/babylon/x/monitor/types"
+	"github.com/babylonlabs-io/babylon/v2/x/monitor/client/cli"
+	"github.com/babylonlabs-io/babylon/v2/x/monitor/keeper"
+	"github.com/babylonlabs-io/babylon/v2/x/monitor/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"