+++ x/zoneconcierge/keeper/epoch_chain_info_indexer.go
@@ -8,8 +8,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/runtime"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	bbn "github.com/babylonlabs-io/babylon/types"
-	"github.com/babylonlabs-io/babylon/x/zoneconcierge/types"
+	bbn "github.com/babylonlabs-io/babylon/v2/types"
+	"github.com/babylonlabs-io/babylon/v2/x/zoneconcierge/types"
 )
 
 // GetEpochChainInfo gets the latest chain info of a given epoch for a given chain ID